+++ ums-api/src/main/java/org/ums/domain/model/mutable/MutableSemesterWithdrawal.java
@@ -11,7 +11,7 @@ import org.ums.domain.model.immutable.Student;
  * Created by My Pc on 3/22/2016.
  */
 public interface MutableSemesterWithdrawal extends SemesterWithdrawal, Mutable,
-    MutableLastModifier, MutableIdentifier<Integer> {
+    MutableLastModifier, MutableIdentifier<Long> {
   void setSemester(final Semester pSemester);
 
   void setStudent(final Student pStudent);
